[SofaImGui] Refactor method Components
