+++ SofaImGui/src/SofaImGui/ImGuiGUIEngine.cpp
@@ -77,7 +77,7 @@
 #include "windows/SceneGraph.h"
 #include "windows/DisplayFlags.h"
 #include "windows/Plugins.h"
-
+#include "windows/Components.h"
 
 using namespace sofa;
 
@@ -278,345 +278,6 @@ void ImGuiGUIEngine::showViewport(sofa::core::sptr<sofa::simulation::Node> groot
     }
 }
 
-void ImGuiGUIEngine::showComponents(const char* const& windowNameComponents, bool& isComponentsWindowOpen)
-{
-    if (isComponentsWindowOpen)
-    {
-        if (ImGui::Begin(windowNameComponents, &isComponentsWindowOpen))
-        {
-            unsigned int nbLoadedComponents = 0;
-            if (ImGui::BeginTable("split", 2, ImGuiTableFlags_BordersOuter | ImGuiTableFlags_Resizable))
-            {
-                ImGui::TableNextRow();
-                ImGui::TableNextColumn();
-
-                static core::ClassEntry::SPtr selectedEntry;
-
-                static std::vector<core::ClassEntry::SPtr> entries;
-                entries.clear();
-                core::ObjectFactory::getInstance()->getAllEntries(entries);
-                nbLoadedComponents = entries.size();
-
-                static ImGuiTableFlags flags = ImGuiTableFlags_ScrollY | ImGuiTableFlags_BordersV | ImGuiTableFlags_BordersOuterH | ImGuiTableFlags_Resizable | ImGuiTableFlags_RowBg | ImGuiTableFlags_NoBordersInBody | ImGuiTableFlags_Sortable;
-                if (ImGui::BeginTable("componentTable", 2, flags, ImVec2(0.f, 400.f)))
-                {
-                    ImGui::TableSetupColumn("Name");
-                    ImGui::TableSetupColumn("Category");
-                    ImGui::TableHeadersRow();
-
-                    struct ComponentEntry
-                    {
-                        std::string name;
-                        std::string category;
-                        core::ClassEntry::SPtr classEntry;
-                    };
-                    static std::vector<ComponentEntry> componentEntries;
-                    if (componentEntries.empty())
-                    {
-                        for (const auto& entry : entries)
-                        {
-                            std::set<std::string> categoriesSet;
-                            for (const auto& [templateInstance, creator] : entry->creatorMap)
-                            {
-                                std::vector<std::string> categories;
-                                core::CategoryLibrary::getCategories(entry->creatorMap.begin()->second->getClass(), categories);
-                                categoriesSet.insert(categories.begin(), categories.end());
-                            }
-                            std::stringstream categorySs;
-                            for (const auto& c : categoriesSet)
-                                categorySs << c << ", ";
-
-                            const std::string categoriesText = categorySs.str().substr(0, categorySs.str().size()-2);
-
-                            componentEntries.push_back({entry->className, categoriesText, entry});
-                        }
-                    }
-
-                    if (ImGuiTableSortSpecs* sorts_specs = ImGui::TableGetSortSpecs())
-                    {
-                        if (sorts_specs->SpecsDirty)
-                        {
-                            std::sort(componentEntries.begin(), componentEntries.end(), [sorts_specs](const ComponentEntry& lhs, const ComponentEntry& rhs)
-                            {
-                                for (int n = 0; n < sorts_specs->SpecsCount; n++)
-                                {
-                                    const ImGuiTableColumnSortSpecs* sort_spec = &sorts_specs->Specs[n];
-                                    const bool ascending = sort_spec->SortDirection == ImGuiSortDirection_Ascending;
-                                    switch (sort_spec->ColumnIndex)
-                                    {
-                                    case 0:
-                                        {
-
-                                            if (lhs.name < rhs.name) return ascending;
-                                            if (lhs.name > rhs.name) return !ascending;
-                                            break;
-                                        }
-                                    case 1:
-                                        {
-                                            if (lhs.category < rhs.category) return ascending;
-                                            if (lhs.category > rhs.category) return !ascending;
-                                            break;
-                                            return lhs.name < rhs.name;
-                                        }
-                                    default:
-                                        IM_ASSERT(0); break;
-                                    }
-                                }
-                                return false;
-                            });
-                        }
-                    }
-
-                    static const std::map<std::string, ImVec4> colorMap = []()
-                    {
-                        std::map<std::string, ImVec4> m;
-                        int i {};
-                        auto categories = core::CategoryLibrary::getCategories();
-                        std::sort(categories.begin(), categories.end(), std::less<std::string>());
-                        for (const auto& cat : categories)
-                        {
-                            ImVec4 color;
-                            color.w = 1.f;
-                            ImGui::ColorConvertHSVtoRGB(i++ / (static_cast<float>(categories.size())-1.f), 0.72f, 1.f, color.x, color.y, color.z);
-                            m[cat] = color;
-                        }
-                        return m;
-                    }();
-
-                    for (const auto& entry : componentEntries)
-                    {
-                        ImGui::TableNextRow();
-                        ImGui::TableNextColumn();
-                        if (ImGui::Selectable(entry.name.c_str(), entry.classEntry == selectedEntry))
-                            selectedEntry = entry.classEntry;
-                        ImGui::TableNextColumn();
-
-                        const auto colorIt = colorMap.find(entry.category);
-                        if (colorIt != colorMap.end())
-                            ImGui::TextColored(colorIt->second, colorIt->first.c_str());
-                        else
-                            ImGui::Text(entry.category.c_str());
-                    }
-                    ImGui::EndTable();
-                }
-
-                ImGui::TableNextColumn();
-
-                if (selectedEntry)
-                {
-                    ImGui::Text("Name: %s", selectedEntry->className.c_str());
-                    ImGui::Spacing();
-                    ImGui::TextDisabled("Description:");
-                    ImGui::TextWrapped(selectedEntry->description.c_str());
-                    ImGui::Spacing();
-
-                    bool hasTemplate = false;
-                    for (const auto& [templateInstance, creator] : selectedEntry->creatorMap)
-                    {
-                        if (hasTemplate |= !templateInstance.empty())
-                            break;
-                    }
-
-                    if (hasTemplate)
-                    {
-                        ImGui::Spacing();
-                        ImGui::TextDisabled("Templates:");
-                        for (const auto& [templateInstance, creator] : selectedEntry->creatorMap)
-                        {
-                            ImGui::BulletText(templateInstance.c_str());
-                        }
-                    }
-
-                    if (!selectedEntry->aliases.empty())
-                    {
-                        ImGui::Spacing();
-                        ImGui::TextDisabled("Aliases:");
-                        for (const auto& alias : selectedEntry->aliases)
-                        {
-                            ImGui::BulletText(alias.c_str());
-                        }
-                    }
-
-                    std::set<std::string> namespaces;
-                    for (const auto& [templateInstance, creator] : selectedEntry->creatorMap)
-                    {
-                        namespaces.insert(creator->getClass()->namespaceName);
-                    }
-                    if (!namespaces.empty())
-                    {
-                        ImGui::Spacing();
-                        ImGui::TextDisabled("Namespaces:");
-                        for (const auto& nm : namespaces)
-                        {
-                            ImGui::BulletText(nm.c_str());
-                        }
-                    }
-
-                    std::set<std::string> parents;
-                    for (const auto& [templateInstance, creator] : selectedEntry->creatorMap)
-                    {
-                        for (const auto& p : creator->getClass()->parents)
-                        {
-                            parents.insert(p->className);
-                        }
-                    }
-                    if (!parents.empty())
-                    {
-                        ImGui::Spacing();
-                        ImGui::TextDisabled("Parents:");
-                        for (const auto& p : parents)
-                        {
-                            ImGui::BulletText(p.c_str());
-                        }
-                    }
-
-                    std::set<std::string> targets;
-                    for (const auto& [templateInstance, creator] : selectedEntry->creatorMap)
-                    {
-                        targets.insert(creator->getTarget());
-                    }
-                    if (!targets.empty())
-                    {
-                        ImGui::Spacing();
-                        ImGui::TextDisabled("Targets:");
-                        for (const auto& t : targets)
-                        {
-                            ImGui::BulletText(t.c_str());
-                        }
-                    }
-
-                    ImGui::Separator();
-
-                    struct DataInfo
-                    {
-                        sofa::type::vector<std::string> templateType;
-                        std::string description;
-                        std::string defaultValue;
-                        std::string type;
-                    };
-
-                    std::map<std::string, std::map<std::string, DataInfo>> allData;
-                    {
-                        const auto tmpNode = core::objectmodel::New<simulation::graph::DAGNode>("tmp");
-                        for (const auto& [templateInstance, creator] : selectedEntry->creatorMap)
-                        {
-                            core::objectmodel::BaseObjectDescription desc;
-                            const auto object = creator->createInstance(tmpNode.get(), &desc);
-                            if (object)
-                            {
-                                for (const auto& data : object->getDataFields())
-                                {
-                                    allData[data->getGroup()][data->getName()].templateType.push_back(templateInstance);
-                                    allData[data->getGroup()][data->getName()].description = data->getHelp();
-                                    allData[data->getGroup()][data->getName()].defaultValue = data->getDefaultValueString();
-                                    allData[data->getGroup()][data->getName()].type = data->getValueTypeString();
-                                }
-                            }
-                        }
-                    }
-
-                    if (!allData.empty())
-                    {
-                        ImGui::Spacing();
-                        ImGui::TextDisabled("Data:");
-
-                        for (const auto& [group, templateData] : allData)
-                        {
-                            const auto groupName = group.empty() ? "Property" : group;
-                            if (ImGui::CollapsingHeader(groupName.c_str()))
-                            {
-                                ImGui::Indent();
-                                for (auto& data : templateData)
-                                {
-                                    if (ImGui::CollapsingHeader(data.first.c_str()))
-                                    {
-                                        ImGui::PushStyleColor(ImGuiCol_Text, ImGui::GetStyle().Colors[ImGuiCol_TextDisabled]);
-                                        ImGui::TextWrapped(data.second.description.c_str());
-                                        const std::string defaultValue = "default value: " + data.second.defaultValue;
-                                        ImGui::TextWrapped(defaultValue.c_str());
-                                        const std::string type = "type: " + data.second.type;
-                                        ImGui::TextWrapped(type.c_str());
-
-                                        ImGui::PopStyleColor();
-                                    }
-                                }
-                                ImGui::Unindent();
-                            }
-                        }
-                    }
-                }
-                else
-                {
-                    ImGui::Text("Select a component");
-                }
-
-                ImGui::EndTable();
-            }
-            ImGui::Text("%d loaded components", nbLoadedComponents);
-
-            if (ImGui::Button(ICON_FA_SAVE" "))
-            {
-                nfdchar_t *outPath;
-                const nfdresult_t result = NFD_SaveDialog(&outPath, nullptr, 0, nullptr, "log.txt");
-                if (result == NFD_OKAY)
-                {
-                    static std::vector<core::ClassEntry::SPtr> entries;
-                    entries.clear();
-                    core::ObjectFactory::getInstance()->getAllEntries(entries);
-
-                    if (!entries.empty())
-                    {
-                        std::ofstream outputFile;
-                        outputFile.open(outPath, std::ios::out);
-
-
-
-                        if (outputFile.is_open())
-                        {
-                            for (const auto& entry : entries)
-                            {
-                                struct EntryProperty
-                                {
-                                    std::set<std::string> categories;
-                                    std::string target;
-                                    bool operator<(const EntryProperty& other) const { return target < other.target && categories < other.categories; }
-                                };
-                                std::set<EntryProperty> entryProperties;
-
-                                for (const auto& [templateInstance, creator] : entry->creatorMap)
-                                {
-                                    EntryProperty property;
-
-                                    std::vector<std::string> categories;
-                                    core::CategoryLibrary::getCategories(entry->creatorMap.begin()->second->getClass(), categories);
-                                    property.categories.insert(categories.begin(), categories.end());
-                                    property.target = creator->getTarget();
-
-                                    entryProperties.insert(property);
-                                }
-
-                                for (const auto& [categories, target] : entryProperties)
-                                {
-                                    outputFile
-                                            << entry->className << ','
-                                            << sofa::helper::join(categories.begin(), categories.end(), ';') << ','
-                                            << target << ','
-                                            << '\n';
-                                }
-                            }
-
-                            outputFile.close();
-                        }
-                    }
-
-
-                }
-
-            }
-        }
-        ImGui::End();
-    }
-}
 
 void ImGuiGUIEngine::showSettings(const char* const& windowNameSettings, bool& isSettingsOpen)
 {
@@ -1075,7 +736,7 @@ void ImGuiGUIEngine::startFrame(sofaglfw::SofaGLFWBaseGUI* baseGUI)
     /***************************************
      * Components window
      **************************************/
-    showComponents(windowNameComponents, isComponentsWindowOpen);
+    Components::showComponents(windowNameComponents, isComponentsWindowOpen);
 
     /***************************************
      * Log window
